requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-appconf==1.0.5
django-bootstrap4==22.3
django-bootstrap5==22.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-imagekit==4.1.0
django-js-asset==2.0.0

django-paypal==2.0
django-social-share==2.3.0
django-taggit==3.1.0
django-tinymce==3.6.0
django-widget-tweaks==1.4.12
idna==3.4
Pillow==8.0.0



python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
